Added Biased Die
A biased die is any type of die which favors one or more values over the others. For example, a die which is biased to 3 has a higher probability of getting 3 as a result. This could be useful as a part of player progression where certain outcomes of an action are more likely to happen based on a biased die. Though like the fair die, this class has a vast number of applications!